Initial commit for web version
This is the initial work-in-progress commit for the web-based version of the Adir tool, currently running locally, soon to be running externally.

This version brings the core logic from the Sheets Adir, including:
- Asset generation-Image generation based on existing Google Ads campaign context or manually uploaded context.
- Asset preview - Review and upload images to Google Ads Asset Library (still in progress)

A new feature has been added:
- Asset removal: Allows users to remove low-performing image assets from their campaigns, according to chosen metrics.
Currently the app serves pmax campaigns, and soon to incoroprate Demand Gen campaigns.

Change-Id: I6684a7e1207e571fffd541b230d3de53029d6c14